+++ src/Plutarch/Extra/Identity.hs
@@ -10,6 +10,7 @@ module Plutarch.Extra.Identity (
 import Generics.SOP (Top)
 import qualified Generics.SOP as SOP
 import Plutarch.Extra.Applicative (PApplicative (ppure), PApply (pliftA2))
+import Plutarch.Extra.Bind (PBind ((#>>=)))
 import Plutarch.Extra.Boring (PBoring (pboring))
 import Plutarch.Extra.Comonad (
     PComonad (pextract),
@@ -96,3 +97,9 @@ instance PApplicative PIdentity where
 -- | @since 1.2.0
 instance (PBoring a) => PBoring (PIdentity a) where
     pboring = ppure # pboring
+
+-- | @since 3.0.1
+instance PBind PIdentity where
+    {-# INLINEABLE (#>>=) #-}
+    xs #>>= f = pmatch xs $ \case
+        PIdentity x -> f # x