+++ src/Plutarch/Extra/State.hs
@@ -17,6 +17,7 @@ module Plutarch.Extra.State (
 import Generics.SOP (Top)
 import qualified Generics.SOP as SOP
 import Plutarch.Extra.Applicative (PApplicative (ppure), PApply (pliftA2))
+import Plutarch.Extra.Bind (PBind ((#>>=)))
 import Plutarch.Extra.Functor (PFunctor (PSubcategory, pfmap))
 import Plutarch.Extra.TermCont (pmatchC)
 
@@ -62,6 +63,14 @@ instance PApply (PState s) where
 instance PApplicative (PState s) where
     ppure =
         phoistAcyclic $ plam $ \x -> pcon . PState $ plam $ \s -> pcon . PPair s $ x
+
+-- | @since 3.0.1
+instance PBind (PState s) where
+    {-# INLINEABLE (#>>=) #-}
+    xs #>>= f = pmatch xs $ \case
+        PState g -> pcon . PState . plam $ \s -> pmatch (g # s) $ \case
+            PPair s' res -> pmatch (f # res) $ \case
+                PState h -> h # s'
 
 {- | Lift a Plutarch lambda into 'PState'.